refactoring test helpers. using .h file now
