+++ native/tests/seal/fracture.cpp
@@ -20,7 +20,7 @@
 #include <unordered_set>
 #include <utility>
 #include "gtest/gtest.h"
-#include "helpers.cpp"
+#include "helpers.h"
 
 using namespace seal;
 using namespace std;
@@ -50,26 +50,6 @@ namespace sealtest::fracture
         //            });
         //            return ret;
         //        }
-
-        TEST(nonNegacyclicNTT, ptxPadding)
-        {
-            std::uint64_t N = 8192;
-            // TODO , test in the evaluator tests..
-            auto all = SetupObjs::New(N);
-            auto ptx = all.random_plaintext();
-
-            all.evaluator.transform_to_positive_ntt_inplace(ptx, 2, all.context.first_parms_id());
-            auto v = plain_to_vector(ptx);
-            // assert correct padding.
-            ASSERT_EQ(v.size(), 4 * 2 * N);
-
-            // asserting the padding is of the following form [p_q1, 0 . p_q2, 0 ..]
-            for (auto i = N; i < 2 * N; ++i)
-            {
-                ASSERT_EQ(v[i], 0);
-                ASSERT_EQ(v[2 * i], 0);
-            }
-        }
     } // namespace regntt
 
     namespace polyval
